Implement controller for creating quote request and sending quote request, Change id to type Integer for user, update MySql table to auto update

+++ backend/Car-Insurance-Portal-Service/src/main/java/com/team10/carinsuranceportalservice/repository/InsuranceAgencyRepository.java
@@ -3,5 +3,5 @@ package com.team10.carinsuranceportalservice.repository;
 import com.team10.carinsuranceportalservice.entity.InsuranceAgency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InsuranceAgencyRepository extends JpaRepository<InsuranceAgency, Long> {
+public interface InsuranceAgencyRepository extends JpaRepository<InsuranceAgency, Integer> {
 }
