+++ backend/Car-Insurance-Portal-Service/src/main/java/com/team10/carinsuranceportalservice/repository/QuoteRequestRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface QuoteRequestRepository extends JpaRepository<QuoteRequest, Long> {
+public interface QuoteRequestRepository extends JpaRepository<QuoteRequest, Integer> {
     List<QuoteRequest> findByUserId(Integer userId);
 }
